requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
crypto==1.4.1
dj-database-url==0.5.0
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Markdown==3.3.4
Naked==0.1.31
Pillow==8.3.1
psycopg2-binary==2.9.1
pycryptodome==3.10.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
shellescape==3.8.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
